GameUtils e MinMax in cpp
