+++ .vscode/settings.json
@@ -60,6 +60,8 @@
         "streambuf": "cpp",
         "thread": "cpp",
         "typeinfo": "cpp",
-        "variant": "cpp"
+        "variant": "cpp",
+        "cstring": "cpp",
+        "map": "cpp"
     }
 }
