+++ cpp/GameUtils.h
@@ -8,13 +8,13 @@
 #include <algorithm>
 #include <sstream>
 #include "MinMax.h"
+#include "GameUtils.h"
+#include "Stats.h"
 #include "Logger.h"
 
 using namespace std;
 
 // COSTANTI E UTILITY
-extern const vector<vector<string>> empty_board;
-extern const vector<pair<int, int>> citadels;
 extern const vector<pair<int, int>> winning_positions;
 
 struct Move {
@@ -28,26 +28,29 @@ Move parseMove(const std::string& move);
 // FUNZIONI DI VALIDAZIONE POSIZIONI
 bool is_within_bounds(int row, int col);
 bool is_castle(int row, int col);
+bool is_citadel(int row, int col);
 bool is_adjacent_to_castle(int row, int col);
-bool is_king_captured(const vector<vector<string>>& board);
-bool is_blocking_cell(int row, int col, const vector<vector<string>>& board);
+bool is_blocking_cell(int row, int col);
+bool is_king_captured(const vector<vector<char>>& board);
 
 // FUNZIONI PER IL TURNOVER
-string get_opposite_turn(const string &turn);
+char get_opposite_turn(const char &turn);
 
 // FUNZIONI DI VALIDAZIONE MOSSE
-bool is_valid_move(Move &move, const vector<vector<string>> &board, const string &color);
+bool is_valid_move(Move &move, const vector<vector<char>> &board, const char &color);
 
 // GENERAZIONE MOSSE
-vector<Move> generate_all_possible_moves(const vector<vector<string>> &board, const string &color);
+vector<Move> generate_all_possible_moves(const vector<vector<char>> &board, const char &color);
 
 // GAME OVER
-std::string is_game_over(const std::vector<std::vector<std::string>>& board);
+std::string is_game_over(const std::vector<std::vector<char>>& board);
 
 // APPLICAZIONE MOSSE
-std::vector<std::vector<std::string>> apply_move(const std::vector<std::vector<std::string>>& board, Move& move);
+std::vector<std::vector<char>> apply_move(const std::vector<std::vector<char>>& board, const Move& move);
+bool is_enemy(const char& enemy, const char& of);
+bool is_allie(const char& allie, const char& of);
 
 // BOARD REPRESENTATION
-std::string format_board(const std::vector<std::vector<std::string>>& board);
+std::string format_board(const std::vector<std::vector<char>>& board);
 
 #endif
