+++ cpp/MinMax.h
@@ -7,20 +7,23 @@
 #include <stdexcept>
 #include <algorithm>
 #include <sstream>
+#include "MinMax.h"
 #include "GameUtils.h"
+#include "Stats.h"
+#include "Logger.h"
 
 
 using namespace std;
 
 // EVALUAZIONE EURISTICA
-int heuristic_white(const vector<vector<int>>& board);
-int calculate_black_starting_positions_points(const vector<vector<string>>& board);
-int heuristic_black(const vector<vector<string>>& board);
-int heuristic(const vector<vector<int>>& board, const string& turn);
-int heuristic_evaluation(const vector<vector<int>>& board, const string& turn, const string& player);
+int heuristic_white(const vector<vector<char>>& board);
+int calculate_black_starting_positions_points(const vector<vector<char>>& board);
+int heuristic_black(const vector<vector<char>>& board);
+int heuristic(const vector<vector<char>>& board, const char& turn);
+int heuristic_evaluation(const vector<vector<char>>& board, const char& turn, const char& player);
 
 // MINIMAX ALPHA BETA PRUNING
-pair<int, Move> minimax_alpha_beta(vector<vector<int>>& board, int depth, int alpha, int beta, const string& turn, const string& player);
+pair<int, Move> minimax_alpha_beta(vector<vector<char>>& board, int depth, int alpha, int beta, const char& turn, const char& player);
 
 
 #endif
